[dev][usb-video] Remove initial call to SetFormat.
I added this in for testing, back when we could not yet choose a format.

There's no point in setting an initial format, as the client needs to
call SetFormatLocked to get a video buffer channel.

This would also cause the video device to be removed if any of the control
requests failed.

Change-Id: Ic33d37385ef00d1c3560debb43e55e4e124d1749