[fidl][go] Don't cleanup until outstanding requests are handled
This change ensures that BindingSet.Remove, BindingSet.Close, and
Binding.Close all wait until outstanding requests are handled before
closing the underlying channel for a binding. This allows one to safely
call "Close" on a Binding that's currently being handled, so dispatch
should always succeed (barring failures to write to the channel).

US-462

Test: CQ, go_fidl_bindings_test

Change-Id: Ibe8934279ed875df27a985edb890220a50f0da0a