[netstack] Implement network reachability API.
OnInterfacesChanged, check the list of interfaces for at least one
interface that is up, has DHCP enabled, and has an address. If there is
one and the network was not reachable before, notify clients that the
network is reachable. If there is not one and the network was reachable
before, notify clients that the network is not reachable.

connectivity.go uses only the public interfaces of netstack to ensure
that we can extract it into a separate service when necessary.

This required exposing a DHCP config flag on the FIDL NetInterface type
so we know that interfaces with addresses were not configured
statically. This will be useful for NET-965 as well.

I've also added conversion methods for netstack.NetAddress <->
tcpip.Address to netstack/fidlconv.
We can remove these later if necesssary.

Tested:
  - fx run -N; fx run-test netstack_netiface_test
  - fx run -N; fx run-test netstack_connectivity_test
  - fx run -N -u scripts/start-dhcp-server.sh
    - change `debug` constant in connectivity.go to true
    - verify that "notifying clients of new reachability status: true"
      is printed once DHCP succeeds.
NET-976 #done

Change-Id: I82f3098defcd927f81e6dfdcda35af9560b88aa6