[audio][mixer] Add pending-init phase for AudioDevices.
Add an explicit phase in the audio device manager for audio devices
which are still pending init.  There are now two sets of audio device
bookkeeping in the device manager; a list of devices which are waiting
to become initialized, and a set of devices which have been
become activated and are ready to do work.

Devices which are waiting to be initialized are not currently
considered for routing purposes, and will not be enumerated via the
up-and-coming device enumeration interface.

Also, remove the ability of AudioObjects to exist in lists.  Now that
the device manager bookkeeping has different flavors (sets for
devices, and lists for capturers/renderers), it makes more sense to
keep audio objects in their own collections with their own appropriate
structure.

Test: Manual testing on NUC with simple_sine and vol.  Checked
      built-in output and USB output.
Change-Id: Ic23d6d519a1345236f836bf952a2775161d4b07c