[guest][tests] Add a basic zircon guest integration test
The test simply launches a zircon guest, waits for a terminal, and runs
an echo command. This will be fleshed out and improved on.

This test has an infrequent panic (on the host), see PD-165.

See INTK-408 for details on getting this test to run on hardware only,
and why we need the workaround to pass the test when the hypervisor is
not supported.

TEST: This is a test.

PD-158 #comment Added zircon_guest integration test.
PD-165 #comment Panic occurs on this test.

Change-Id: Iae29d89b08e2edcde076659311384b025f815f8f