[pkg] Packageize modular_tests.
Move run_modular_tests.sh off the deprecated system package.

Shell scripts are not well supported for execution as a
package yet. Specifically, the shell script does not run
in the context of its package. Therfore, it sees the
config file from its own package under its full path
name, as does the binary it's passed to, because that's
still in the system image.

MI4-1018 #comment Packageize modular_tests.

Change-Id: Ib28a6681e2d06a8018d531704a383a773dd8f70e