[amlogic][rawnand] Add a TODO: for tuning of the controller delay.
It is set to 200us now, but testing different values for delay,
a delay as low as 20us works. Don't want to change this right now
before testing various combinations of controller/NAND that we need
to support. Also, this should be a controller specific parameter,
instead of a hardcoded constant that applies to all controllers.

Change-Id: Ie20a68814a81c1be5f364aac4b384ee3e77fd81b