[wlan][mlme] Adds frame specialization
Rather than taking a Frame's Packet and turning it into a different
more or less concrete Frame, WithBody<Body> should be used
instead. Using the former approach can cause buggy code when working
with advanced frames, advanced as in "Frames use offsets".

This will allow to ask take() to only be used when the Frame is no
longer needed and only its underlying data is of interest.

Test=Included

NET-902 NET-912

Change-Id: I35a3eca7cf071b0e8616c87dd23b3c4857dd97e2