[debugger] Implement step-into
"step" is implemented by single-stepping the CPU until the instruction
pointer is on an instruction associated with a different source line.
The behavior is still pretty buggy as the source line gets lost. I plan
to do more testing and fix those bugs in followups.

Adds debug_agent support for stepping machine instructions automatically
within a range of addresses.

Adds a symbol query function to retrieve the detailed line information
for a given address. This is how the range is computed for giving to the
backend. Some new primitives are added in the client to deal with
ranges of addresses.

Adds plumbing for the console commands.

Fix some errors in line handling identified by unit tests.

TEST=unit

Change-Id: I05e342c9e521d8bab54c4f3a882ee7372b188ef9