[fidl][go] Pool message buffers
This change pools message buffers which mitigates the issues in F5-47.
It's also more effecient and desirable in general. Despite the
appearance of a stack allocation, the buffers for FIDL messages were
getting passed directly into a virtual method call, which resulted in
their forced allocation, despite the comment stating that these arrays
were allocated on the stack.

F5-47

Change-Id: Ie75aab2eecf5d6e29650deab4007dbab1f1ded9d