[kernel][timer] Use specialized scheduler preemption timer
Prior to this change, each CPU's preemption timer was implemented as a
regular old kernel timer. This was fine except that it made performing
graceful shutdown difficult.

During graceful shutdown, as we shutdown each secondary CPU, we move
its timers onto the primary CPU. However, the preemption timer is a
special case and should not be migrated. The CPU that's shutting down
continues to need it until the very end.

This change:

- creates a dedicated, non-migratable, per CPU preemption timer
- makes the preemption timer more efficient (no more locking)
- fixes a panic during shutdown timer migration (ZX-2250)
- is a prerequisite for ZX-2228
- collapses some initialization logic in init.cpp

Tested: Did the following in addition to CQ tests

- (arm64 and x64) ran "k ut all" in a loop
- (arm64 and x64) patched in a change that migrates timers on
  shutdown, repeatedly booted/shutdown for several hours
- (arm64) imaged VIM2 with a zedboot based on the patched kernel above
  to verify mexec path works

Observed no hangs, no panics.

ZX-2250 #done

ZX-2228 #comment prereq

Change-Id: I65485b41a37130b9f116f5d3a92e28e5d681b571