[symbolize] Decompose presentor using new pipeline abstraction
This change introduces a new pipeline abstraction that allows
many parts of the pipeline to be seperated allowing for a minimum
of concerns on the pipeline code itself. This change uses it to
improve the presentor but I plan to use this for Filter, Demuxer,
a parser, etc... For instance Filter will be split into a line
parser and an annotator. Demuxer will remain more or less the
same. Many features such logging and keeping track of certian
error states can be added just as parts of the pipeline.
For instance 'modNamesByBuildID' exists just to keep track of
a certain kind of error. This can just be its own node in the
pipeline and can then collect data more neatly. This will also
make adding line-splitting/merging quite easy in the future.

Change-Id: If2d7ee6245dbd7e3c778d9a9187e94c7440b4715