[mediaplayer] prevent task running on deleted ReaderCache
When a player's demux shuts down, the ReaderCache and Reader that
feed the demux are deleted. The ReaderCache owns the Reader, which
calls back into the ReaderCache. In some cases, the Reader was
calling back into a deleted ReaderCache. This was happening when
making the player switch to a new source when the old source was
still playing.

This CL fixes the problem using a weak_ptr to make sure the
ReaderCache still exists when the callback is invoked. This involves
making the ReaderCache inherit std::enable_shared_from_this and
ensuring that a shared pointer to the ReaderCache has been created
before shared_from_this is called. This, in turn, entailed making
sure that a reader's Describe method callback never calls back
synchronously. The Incident utility class was enhanced to make this
easier.

Change-Id: I7b8c3e9e4015f8be038c3aeaf2576daa2c7325c7