[debugger] Printing improvements
Fix an assertion failure when printing stack frames but no stack frames
are available. The code now prints out "no stack frames".

Cleans up printing of location information when a thead stops.

Adds function names to locations (appears in thread stops, frames, etc.).

Separates out the arrow character I use for marking the current code
line and use it for the other places we show the current row. This is
not a constant so we can provide a dynamically selected non-Unicode one
in the future if necessary.

Change-Id: I8d1f5e1dd77d043871db11381f7fe42454d88d6e