[debugger] Improve indexing memory usage.
The old code stored compile unit pointers in the index which meant we
had to keep those objects around, even though they hold a lot of memory.

This change uses compilation unit indices in the index which are
reconverted back into parsed DWARFCompileUnits as needed. This lets us
free all of the compilation units used by the index and saves a lot of
memory for large programs.

Change-Id: I07457daa4238d14dcd6fb5e5b4a7fbfdd076acf5