[mkzedboot] Use a file descriptor to lock device
On systems with an automounting daemon, mkzedboot could fail due to EBUSY and
other errors returned by the host operating system due to races with the
automounter. On those systems the event to rescan the partition table, thus
causing automounting occurs when the raw disk device has all open file
descriptors closed. As such, this patch holds open a file descriptor until
all operations that modify the disk are completed.

Change-Id: Ifcab97918bc5b8d6e191e096fbe5ecaa132bee6d