[wlan][mlme] Various small improvements to Frame
(1) Rename take() to Take()
(2) Remove some if-stmts which followed debug asserts:
Originally these statements where intended to handle error cases in
release builds gracefully. However, the caller of such methods is
not expecting these methods to ever return a nullptr and never checks
that case. Hence, all these if-stmts do is to move the crash to another
location. Instead, I prefer having the system crash always in the same
place which will help with understanding problems.
(3) Allow things such as "len()" and "body_len()" to be called on taken
frames.

Test=Existing unittests cover these changes

NET-902 NET-912

Change-Id: I17c47179be7f644f2f96a1e2028e628d20211e19