[amlogic-video] Add support for not using the ES parser
For elementary stream video input the parser seems to mainly copy the
video, switch the endianness around, and update some fifo registers.
For clear content this isn't too useful, so add a mode where the video
is copied directly into the stream buffer and the fifo registers are
updated manually.

This will be particularly useful for multi-instance video, as it can
hopefully avoid context switching the parser registers.

TEST: amlogic_decoder_tests=true

Change-Id: I97f96e2cf82710b7d1db736cccd1f61502e5bb58