[dev][intel-hda][dsp] keep FW pinned until it is loaded
The FW VMO was getting unpinned while the code loader DMA
was in progress, because the PinnedVmo object was destroyed
on function return from IntelDspCodeLoader::TransferFirmware()
which is not responsible for DMA completion wait.

Move the PinnedVmo object to the function that waits for
firmware load success (implies DMA completion), so the memory
stays pinned.

Change-Id: I207a7ca6a68a448011eea6829251c2040d29d986