[ulib][tftp] Fix invalid references to containing function constexprs from lambda
Even constants that are actually just integrated at compile time
technically need to be in a capture to be visible in a lambda when
defined in a containing function scope.  A lambda with captures
cannot be converted into a function pointer, so captures can't be
used here.  Instead, the declarations need to be in a containing
scope with permanent extent.

Change-Id: I24474acd1b4bd42736f77c36394cea34bb0d9f3c