[prebuilt] Update to GCC 8
This compiler is the stable 8 branch, which now contains all our fixes.

It has new warnings that trigger in many places in our code,
disabled for now by -Wno-format-truncation pending investigation.

The for loop in VmAddressRegion::PageFault triggered a "might
be used uninitialized" warning that is probably a compiler bug.
But the equivalent while loop doesn't get the warning.

Change-Id: I6d0bba067d4f8918d19644a068f2d915ce6fc9ad