[Codec] IOCTL def, to create/get CodecFactory
This IOCTL will be used to get a CodecFactory FIDL channel established
to drivers like amlogic_video_decoder.  The IOCTL asks for the client
end of a CodecFactory interface - the driver will create the channel and
bind to the server end.

This change is needed in zircon before corresponding usage in garnet.

Tested: built a change in garnet that uses the defined IOCTL name
Change-Id: I319ee78df3fc4ffa517f9534b072699ec38b26a6