[debugger] Improve context printing.
Enhances line printing to better integrate source and disassembly views.
It now falls back to disassembly when a thread stops and there are no
symbols. Since disassembly is asynchronous, some new infrastructure in
the context formatter is added to completely manage the dumping of
context.

Disassembly formatting is now moved from the disassembly verb to the
context formatter so it can be shared. Most of the options are removed
from the disassembler and it outputs a structure that the formatter
interprets according to the UI requirements. This allows for proper
handling of addresses (marking the active address).

The code formatter now knows how to display breakpoints in the margin of
a code listing. This is not enabled currently for listings, but when you
set a breakpoint the location of the breakpoint with a small amount of
context is added which helps verify that the operation did what you
wanted.

Fixed address-based breakpoints which were broken. The address was not
getting registered. A unit test is added for this case.

TEST=unit

Change-Id: I97befe5bfcb0afb2fea0e7e4d696b7d49dcf041f