[fx] -i: use inotifywait monitor and batch results
Before this change, this sequence of events was possible:
$ fx -i my-command
...
File 1 modified
Command my-command is run
File 2 modified
Command my-command terminates
(Second modification never triggered another my-command run)

With this change, any number of modifications captured after a command
was triggered but before it terminated will result in another command
trigger.

In addition, previously we'd monitor for changes only while waiting to
run a command, and we'd re-establish monitors afterwards, which exposed
the tool to a short gap during which changes were not monitored. Now we
monitor continuously.

Change-Id: I7078283e3d009f23a2dcc0261ef7eef45eadb320