[garnet][timezone] Add debug statement.
This adds an error statement that will assist with SY-516.
Coincidentally, this also appears to cause the errors in reporting the
correct offset to stop.  If they pop up again, this will at least speed
up hunting down the issue.

SY-516 #comment This appears to prevent loading issues as well. Unclear.

TEST=manual on-device testing in debug and release build using tz-util

Change-Id: I26f95062340dec379ea7ac169f0482c334ffec88