[acpica] Set alignment on acpi_table_facs to appease Clang
The acpi_table_facs, like many other ACPI structs, uses byte-packing to
match ACPI specification, but that means Clang insists on generating a
call to __atomic_compare_exchange_n intrisic to satistfy alignment
requirements when trying to atomically set the value for GlobalLock.
This is undesirable, especially since acpi_table_facs does not need
one byte alignment. So to appease Clang, we explicitly set the 4 byte
alignment for this struct, and enable -Wpadded error for this struct to
ensure we catch any potential future regressions.

Change-Id: I638bd500ba01d5a6f38ccd598e66a6ba65460e84