[debugger] Use build dir as root for symbol paths.
Exposes the build directory (which was already used to find ids.txt) and
uses it as a search path when looking for source files.

Normally GDB can find these source files because it's looking relative
to the binary. But since this is a remote debugger, that has no meaning,
and we need to use the build directory instead.

Deletes some _backup.* files I accidentally checked in.

TEST=unit, manual

Change-Id: Icf8d617d585bdf358190fd524e9ba49d08d990c8