[garnet][timezone] Initialize error code.
Due to the implementation of `getOffset()`, the `UErrorCode` variable
needs to be initialized to `U_ZERO_ERROR` in order to function
correctly, else the function will return zero (this is due to a check
for a failure code at the beginning of `getOffset()`)

Also contains retroactive-review fixes.

TEST=manual verification on machine w/ reproducible failures from prev patch

SY-516 #comment This change should fix the random failure issue.

Change-Id: Ib163a28a9f5f54ba8a7ed1a911b046ef0ac71b95