[debugger] Fix assertion on breakpoint disable/endable
Previously if a breakpoint was disabled, the client will notify the
backend that the breakpoint was and the Breakpoint object will be
deleted. But the individual BreakpointLocations were not updated which
would result in a crash up re-enable.

Notify the indivual locations when the breakpoint is removed, and add a
test that these notifications happen.

TEST=unit

Change-Id: I4fbafce3abb4237b4493c7de6c41afe38d08e373