[media][examples] Make simple_sine simple again
Over time the simple_sine example has accreted a number of features,
as the example was used as a handy tool for testing. However, since
this example is intended to be used as the simplest-possible example
for streaming audio using the FIDL interfaces, it will be most
effective if it is slimmed back down to its original purpose.

The newly-added tool signal_generator has flexibility (via cmd-line
parameters) for everything that one might have previously wanted to
do with simple_sine.

Test: manually ran the example on NUC and ARM64 QEMU.

Change-Id: I5055a648aacc5a79af5bfe3eecbfb9ec62ac37dd