[vmm] Fix zircon BOOTFS entry in ZBI
Until now we have been reporting the size of the BOOTFS entry in the
guest ZBI by including the size of the container header (which we
strip). This doesn't seem to have been a problem until some recent
changes to the ZBI library. Without this patch the guest attempts to
read an entry past the end of the ZBI and fails.

TEST: Manually tested by booting zircon_guest with display=none and
display=scenic.

Change-Id: I8cbedaed8c41d4f946c62c82b2f57686d335246e