[examples/ui] Fix typo in tile_view so it can provide env services
The code was attempting to wire up environment services to its child
realm in the error case, not the success case. This mean environment
services weren't available to anything created in the tile realm.

This appears to be a typo introduced by a mechanical code refactor that
omitted a semicolon after a 'return':
https://fuchsia.googlesource.com/garnet/+/96ea9d9ce8fefe4fc0cfae41c0e4e672e6f45d57%5E%21/#F12

which was later reformatted to the current code.

Test: Run 'set_root_view tile_view shapes_view' and see some shapes

Change-Id: I838cc306875f524e47a51a7cdf1ead3527c6530b