[zircon][sdio] Add support to get oob irq handle
This changeset adds support to get the out-of-bandwidth irq handle for
SDIO. The irq needs to be accessed by the upper layers in sdio driver
stack but the gpio information is at the board level. In future we
need to think of a way to access the interrupt directly in the upper
layer without the involvement of the controller.

Test: Sanity check of EMMC on pixelbook, VIM2, sdio on estelle

Change-Id: I39919f79e006e598172db91f6d02855b8986f816