[Codec] use_h264_decoder() toward h264 decode.
This parses a .h264 (00 00 00 01 start codes) file for NALs, sends those
NALs into a video/h264 Codec that it got from amlogic_video_decoder
driver, and will dump the output frames to a raw YUV frame file using
RawVideoWriter once the decode is hooked up.

The use_aac_decoder executable will be renamed later after the
use_h264_decoder() function has decoded some h264, to reflect that the
program is not only to decode AAC.  For now, the name of the program is
still fine as-is.

Tested: use_aac_decoder_test still works
Change-Id: Iea1feb97e413aaeb3944eca7fd04e14c20f21759