[suggestion_engine] Add support for rich suggestions
This adds support for rich suggestions to the suggestion engine.

If a proposal is added with wants_rich_suggestion = true, the
suggestion engine will create a new story and perform the associated
actions prior to adding the proposal to the next processor.

This will allow the component responsible for displaying suggestions to
display the created story as a rich suggestion.

Note: This change was previously submitted and reverted due to the test
being flaky. This registers an activity with the idle waiter so the
WaitUntilIdle() in the test will do the right thing.

TESTED=suggestion_engine_test

Change-Id: Ib8b2ae5486eafa74ccd31d9769119545328bb36f