[suggestion_engine] Add support for rich suggestions
This adds support for rich suggestions to the suggestion engine.

If a proposal is added with wants_rich_suggestion = true, the
suggestion engine will create a new story and perform the associated
actions prior to adding the proposal to the next processor.

This will allow the component responsible for displaying suggestions to
display the created story as a rich suggestion.

TESTED=suggestion_engine_test

Change-Id: Id544c68c63af40fd0cebd81014b2ce5ab9e035b5