[paver] migrate to new partition layout on cros
Alter the partition layout on CROS to include ZIRCON-A, ZIRCON-B and
ZIRCON-R. At this stage only ZIRCON-A will be used, but the others will
begin to be used in following changes.

Test: CQ: chromeos-disk-setup, plus manual plan below.

* Tested against a freshly "recovered" device
* Tested aginst a device paved with the prior generation paving code
* Tested against itself, repeated paving

In the manual tests above, checked `cgpt show` to verify starting state,
and inspected `lsblk` for the correct resultant state.

Change-Id: I1faa76cb2689198db84612a947d224c3ee938e30