[debugger] Improve location input
The breakpoint location information was previously part of
BreakpointSettings but this concept needs to be used for things like
"until" or "list". The parsing code is basically unchanged but in a new
file and with updated names. The documentation is split out in a way
that can be re-used by the various function's help.

Adds a no-argument version of "break" to set a breakpoint at the current
frame's location. Better documentation of locations.

TEST=unit

Change-Id: Ic6fce27bf3743194ce8954bc11ad1dd5786d1e98