[zxcrypt] Zero-copy reads
Relanding CL 156789 with flake fixes in place.

This CL refactors the zxcrypt encrypted block device filter driver. this
driver uses an internal buffer to hide cryptographic transformations
from the upper layers.  For example, if a filesystem uses some memory to
write a block and then reads from that memory, it may reasonably expect
to find the plaintext data it wrote.

However, in the case of reads the previous contents of the buffer are
always clobbered.  As a result, there's no downside to doing the
decryption in-place, and the internal buffer can be avoided.

As part of this refactor, efforts were also made to lean more on atomics
where possible and reduce locking to only those concerns which directly
affect the internal write buffer.  As a consequence, the lifecycle code
around DdkUnbind and DdkRelease was also improved.

ZX-1811 #comment Bypasses copying between VMOs on read.

Test: zxcrypt unit tests.
Change-Id: I9755a109d3ed509f95b1f40eab5e2a896cfa6bd9