[tee] Add TEE Client Application for testing
As the TEE interface is being developed, it's useful to have a test app
that attempts to connect to its Trusted App partner in the secure world.
Since we don't yet have a Trusted App, this test can do nothing more
than initialize a context with the TEE.

Test: Manually executed tee-test on device and validated that it
returned a NOT_IMPLEMENTED error code.
Change-Id: Ica2cc2f7e89b95c2bcea07f7d19245b0d98b9ffd