[tee][optee] Add Context initialize and finalize
The first building block concept for the TEE Client API is a Context.
A Context is a logical connection which exists between a Client
Application and a TEE. For our implementation, this is nothing more than
opening and closing the TEE device driver. As a precaution, we also
validate that the TEE device we've opened is compliant with
GlobalPlatform.

Test: Manually validated that the tee-test succeeded in opening the
device.
Change-Id: I3ea929f8148069c490b6eba83c28e5fa2808b441