[sysmgr] Add ability to start services up eagerly in addition to apps
This teaches sysmgr how to start services up eagerly in addition to
launching apps eagerly based on the service configuration data
determined at build time.  This allows deleting the code that hardcodes
the names of the netstack and wlan services in sysmgr (along with their
odd test configuration check) and removing the amber_init program which
was a startup app that existed purely to provide this capability.

Test: service_startup_test, verify that netstack, wlan, and amber start
eagerly in a full product build

Change-Id: Icc12ffebbc8103733bed226af67c041387a8a8cb