[zircon][sdmmc] Avoid passing down the block op to controllers
aml-sd-emmc controller and possibly sdhci will be used for sdio
transactions as well. This means that the requests to the controller
may come either from block drivers or from a driver in the stack of sdio
devices (For ex: wifi).This changeset avoids passing down the block op
data structure down to the controllers for setting up the dma descs.
Instead, for a dma data transaction, we pass on the vmo, vmo offset that
we got from the block op.

Test: A quick dd test on EMMC on vim2 and Eve. I will follow it up with
iochk while review is in progress

Change-Id: I899aa7f32a39b464fe9fd0a33fa5d386b719cc68