[fidl] Add fidl_msg_t and fidl_txn_t
These types will be used by the simple C bindings.

Originally, I was going to put these in lib/fidl/somethinng.h, but that
introduces a header-only dependency from the C header for each FIDL
library to libfidl. It's pretty cheap to put this definition in
zircon/fidl.h, which avoids that dependency.

Test: No behavior change
Change-Id: I20f266a0c492365f24f0c8e3972b021bd88feb7d