[rust] Fix depfile generation
When root_out_dir doesn't match root_build_dir, the rust depfile has
the wrong target and so ninja will always rebuild the target.  The
target path in the depfile needs to be relative to root_build_dir,
which is the cwd.

SEC-27 #comment Fix gratuitous incremental rebuilds of rust in variants

Change-Id: Ie4684f5c848e4bac73816909fc437eb4b26609d3