[fidl] Start migration from bool to zx_status_t
This CL starts the process of converting FooSyncPtr methods from
returning bool to returning zx_status_t. This conversion is tricky
because zx_status_t implicitly converts to bool with the opposite sense
in which FooSyncPtr uses bool.

This CL introduces FooSync2Ptr, which returns a struct that contains the
zx_status_t. Future CLs will migrate clients from FooSyncPtr to
FooSync2Ptr and use the fact that the struct will cause a compile error
when implicitly converted to bool to find all boolean conversions.

After that, we'll change FooSyncPtr to return zx_status_t and migrate
all the clients back.

FIDL-223 #comment

Test: No behavior change
Change-Id: I1659f7793b35ccb3af2530b99b0d982b8faf3893