[zircon][sdio] Publish sdio child device
During the sdmmc bind function we add a block device. After the probe
is done when we know the device is an SDIO, we remove the block device
and an SDIO device with manufacturer id and product id as the props for
the upper layer to bind.

Test: SDIO probe on VIM2, astro. Wrote a basic test driver and ensured that published
device is able to bind to it.

Change-Id: I3fc14a9975ccc2e36ca040965d8c48d7b9439b3f