[debugger] Hook up breakpoint notifications.
Adds the notion of an internal breakpoint with a controller object that
can filter what happens when the breakpoint is hit. This will allow
higher-level filtering of whether a breakpoint should "count" from the
client side.

Add the breakpoints that caused a thread stop to the notification. This
will be used in the future to show more context about why the program
stopped and to set the current breakpoint context.

Make the backend breakpoint IDs more stable (previously BreakpointImpl
would use whether this was 0 or not to indicate whether it was
installed). Create a map of these in the SystemImpl object so it can
easily find which breakpoints the debug agent is reporting.

Thread stop notifications now happen in two phases: First the states of
everything are updated, and then callbacks are issued. This ensures that
the state of everything is up-to-date when the callbacks run.

Separated out the RemoteAPI-style of testing used in BreakpointImpl
tests to a helper class. Share this with the new Session tests.

TEST=unit

Change-Id: I780913177fca84e281e3e5e363191eee4fe1f9c6