[async][runtests] Specify some runtests tests as 'medium'.
There is evidence that some of the RunAllTests() host tests fire the
watchdog timer on the bots:
https://logs.chromium.org/v/?s=fuchsia%2Fbuildbucket%2Fcr-buildbucket.appspot.com%2F8942983877130324176%2F%2B%2Fsteps%2Fhost_test_results%2F0%2Fsteps%2Fs___zircon_build-x64_host_tests_runtests-utils-test%2F0%2Flogs%2Fstdio%2F0

For safe measure, I mark all RunAllTests() and RunTestsInDir() here
as MEDIUM.

Tested: this change can only make *more* tests pass. We must necessarily
wait until this change is merged and the test runs in CQ before we can
verify whether host runtests-utils-test successfully runs on the bots.

Change-Id: I658521f8831b81e4f200ed11ed437168812ca9d1