[ledger] Allow flatbuffer messages to be truely movable.
Flatbuffers are usually not movable and passed by pointer. This prevents
us to use flatbuffer objects in an asynchronous context, where the
original storage no longer exists.

We create here a MessageHolder class to hold the flatbuffer data and its parsed
representation so we can move it.

TESTED=ledger_unittests,ledger_integration_tests

Change-Id: I8e64adac21d3f87fdb1b4e549e50ca3af4376e8b