[ledger] Move tests off the deprecated system image
As part of the move, we split benchmarks into their own package
`ledger_benchmarks`, where we bundle all benchmark binaries and tspec
files. As a result, benchmarks are not runnable directly from the shell,
but are available in /pkgfs/packages/ledger_benchmarks/0/bin and tspec
files need to use this absolute path.

Also fix package definition to make sure runbenchmark from garnet is
imported since it is required to run the wrapper script.

LE-495

Tested:
  fx shell "/pkgfs/packages/peridot_benchmarks/0/bin/benchmarks.sh /tmp"
  fx shell "/pkgfs/packages/ledger_tests/0/test/ledger_unittests"
  (and all other tests in that directory, including disabled/ ones)
Change-Id: I02e34b2dd65f9b115c9590812290d71cc06e0a79