[rust] Move allocator setting out of fuchsia-zircon
fuchsia-zircon can appear in the build twice because it is
depended on by third-party libraries, so it is build both as part
of the first-party build and as a third-party build.

When the global allocator was specified in fuchsia-zircon,
this caused a conflict. This change resolves that conflict by moving
the global allocator specification into its own crate.

It also makes it easy for many binaries to automatically use the
system global allocator by pulling it in as a dependency of
fuchsia-async.

Test: built successfully

Change-Id: Ib90f4052c8792a14e2735c9201fe0ca0255c89e4