[display] Implement new vsync API
This change adds a new vsync event to the display controller API removes
the present signal event. It also changes the behavior of the cursor
position API to clamp the cursor to the display, to sidestep the issue
of how vsync interacts with a cursor that isn't on the display.

Test: display-test, vkcube, vkcube_image_pipe_swapchain
Change-Id: I6f332e3709eef2c6e1c932581024c2149b65a2a1