[dev][aml-gpu] Fix setting clock first time.
Switching the dynamic mux between a disabled clock and an enabled clock
doesn't work, so the GPU wouldn't start and "dm reboot" would hang
waiting for the driver, which was hung on a register read.

This happened to work when zedbooting because the zedbooted kernel would
switch the clock a second time, cause it to start working.

TEST: "dm reboot" and system/bin/vkcube with a paved vim2.
ZX-2298 #done

Change-Id: Ic8a338a9bf38c91af3d4c30f7b69b77d377690e3